Move stores to src folder
